Address initial peer review comments
