+++ geometric_kernels/lab_extras/jax/extras.py
@@ -9,10 +9,12 @@ _Numeric = Union[B.Number, B.JAXNumeric]
 
 
 @dispatch
-def take_along_axis(a: _Numeric, index: _Numeric, axis: int = 0) -> _Numeric:  # type: ignore
+def take_along_axis(a: Union[_Numeric, B.Numeric], index: _Numeric, axis: int = 0) -> _Numeric:  # type: ignore
     """
     Gathers elements of `a` along `axis` at `index` locations.
     """
+    if not isinstance(a, jnp.ndarray):
+        a = jnp.array(a)
     return jnp.take_along_axis(a, index, axis=axis)
 
 
@@ -43,7 +45,7 @@ def logspace(start: B.JAXNumeric, stop: _Numeric, num: int = 50):  # type: ignor
 @dispatch
 def degree(a: B.JAXNumeric):  # type: ignore
     """
-    Diagonal matrix with x as main diagonal.
+    Given a vector a, return a diagonal matrix with a as main diagonal.
     """
     degrees = a.sum(axis=0)  # type: ignore
     return jnp.diag(degrees)
