+++ geometric_kernels/lab_extras/tensorflow/extras.py
@@ -53,7 +53,7 @@ def logspace(start: _Numeric, stop: _Numeric, num: int = 50, base: _Numeric = 50
 @dispatch
 def degree(a: B.TFNumeric):  # type: ignore
     """
-    Diagonal matrix with x as main diagonal.
+    Given a vector a, return a diagonal matrix with a as main diagonal.
     """
     degrees = tf.reduce_sum(a, axis=0)  # type: ignore
     return tf.linalg.diag(degrees)
