+++ geometric_kernels/spaces/base.py
@@ -5,6 +5,7 @@ import abc
 
 import lab as B
 
+from geometric_kernels.lab_extras import take_along_axis
 from geometric_kernels.spaces.eigenfunctions import Eigenfunctions
 
 
@@ -50,3 +51,33 @@ class DiscreteSpectrumSpace(Space):
         :return: [num, 1] array containing the eigenvalues
         """
         raise NotImplementedError
+
+
+class ConvertEigenvectorsToEigenfunctions(Eigenfunctions):
+    """
+    Converts the array of eigenvectors to callable objects,
+    where inputs are given by the indices. Based on
+    from geometric_kernels.spaces.mesh import ConvertEigenvectorsToEigenfunctions.
+    """
+
+    def __init__(self, eigenvectors: B.Numeric):
+        """
+        :param eigenvectors: [Nv, M]
+        """
+        self.eigenvectors = eigenvectors
+
+    def __call__(self, X: B.Numeric, **parameters) -> B.Numeric:
+        """
+        Selects `N` locations from the `M` eigenvectors.
+
+        :param X: indices [N, 1]
+        :param parameters: unused
+        :return: [N, M]
+        """
+        indices = B.cast(B.dtype_int(X), X)
+        Phi = take_along_axis(self.eigenvectors, indices, axis=0)
+        return Phi
+
+    def num_eigenfunctions(self) -> int:
+        """Number of eigenvectors, M"""
+        return B.shape(self.eigenvectors)[-1]
